+++ changelog.md
@@ -26,6 +26,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Nothing.
 
+## 0.2.0-preview - 2019-11-19
+- [Commits](https://github.com/unity-game-framework/ugf-module-scenes/compare/0.1.0-preview...0.2.0-preview)
+- [Milestone](https://github.com/unity-game-framework/ugf-module-scenes/milestone/2?closed=1)
+
+### Changed
+- Package dependencies:
+    - `com.ugf.module`: from `0.1.0-preview` to `2.0.0-preview`.
+- Rework module to work with async / await instead of coroutines.
+
+### Deprecated
+- Nothing.
+
+### Removed
+- Package dependencies:
+    - `com.ugf.coroutines`: from `0.1.0-preview`.
+
 ## 0.1.0-preview - 2019-10-08
 - [Commits](https://github.com/unity-game-framework/ugf-module-scenes/compare/09b9c96...0.1.0-preview)
 - [Milestone](https://github.com/unity-game-framework/ugf-module-scenes/milestone/1?closed=1)